requirements.txt
joblib==1.2.0
matplotlib==3.5.2
numpy==1.24.3
opencv_python==4.7.0.72
Pillow==9.5.0
plotly==5.14.1
scikit_image==0.19.3
scikit_learn==1.2.2
skimage==0.0
streamlit==1.20.0
streamlit_option_menu==0.3.2
tensorflow==2.12.0
tensorflow_intel==2.11.0
